+++ projects/api/src/contracts/users/_internal/response/movieActivityHistoryResponseSchema.ts
@@ -1,9 +1,9 @@
 import { movieResponseSchema } from '../../../_internal/response/movieResponseSchema.ts';
-import { z } from '../../../_internal/z.ts';
+import { int64, z } from '../../../_internal/z.ts';
 import { historyActionSchema } from './historyActionSchema.ts';
 
 export const movieActivityHistoryResponseSchema = z.object({
-  id: z.number().int(),
+  id: int64(z.number().int()),
   watched_at: z.string().datetime(),
   movie: movieResponseSchema,
   action: historyActionSchema,
